refactor: Break out farm fetch into own helper
